Grid has been split up into a seperate header and source file, definitions in the header and implementations in the source.
Grid also now has a constructor so creation of a grid object is now
slightly different, message me if you can't figure it out.

Coordinate has it's own file now.

Other changes are mostly temporary hacks to get this working until I can
get around to fixing the algorithms properly.

High possibility of bugs, if you feel capable of fixing them please feel
free, or just message me with the problem and I'll deal with it as soon
as possible